refactor : modal onclick -> modalHandler로 이름 변경
