+++ src/components/@common/GroupColorList/index.tsx
@@ -3,17 +3,17 @@ import { Dispatch, SetStateAction } from 'react';
 import { COLORS } from '@/constants/Group';
 import { GroupColor } from '@/types/group';
 
-interface GroupColorList {
-  value: string;
+interface Props {
+  selectedColor: GroupColor;
   onChange: Dispatch<SetStateAction<GroupColor>>;
 }
 
-export const GroupColorList = ({ value, onChange }: GroupColorList) => {
+export const GroupColorList = ({ selectedColor, onChange }: Props) => {
   return (
     <Style.GroupColorList onClick={(e) => e.preventDefault()}>
       {COLORS.map((color) => {
         return (
-          <Style.SelectedButton select={value === color} key={color}>
+          <Style.SelectedButton select={selectedColor === color} key={color}>
             <Style.ColorButton color={color} onClick={() => onChange(color)} />
           </Style.SelectedButton>
         );
