+++ src/components/@common/GroupColorList/styles.ts
