+++ src/components/@common/index.tsx
@@ -3,7 +3,6 @@ export { default as Calendar } from './Calendar';
 export { default as DropBox } from './DropBox';
 export { default as DropDown } from './DropDown';
 export { default as Input } from './Input';
-export { default as InviteModal } from './InviteModal';
 export { default as KakaoLogin } from './KakaoLogin';
 export { default as Label } from './Label';
 export { default as Modal } from './Modal';
