+++ src/components/Home/Header/UserConfig/index.tsx
@@ -55,7 +55,7 @@ const UserConfig = () => {
         <TwoButtonModal
           title="로그아웃"
           description="로그아웃 하시겠습니까?"
-          onClick={handleShowLogOutmodal}
+          modalHandler={handleShowLogOutmodal}
           cancel={{ text: '취소', onClick: handleShowLogOutmodal }}
           confirm={{ text: '로그아웃', onClick: onClickLogOut }}
         />
